+++ packages/terra-table/tests/wdio/table-spec.js
@@ -111,14 +111,14 @@ Terra.describeViewports('Table', ['medium', 'large'], () => {
       browser.url('/raw/tests/cerner-terra-framework-docs/table/table-focusable-cell');
     });
 
-    it('validates that the proper grid element is selected when Tab is used to give initial focus to the grid', () => {
+    it('validates that the first interactive element in the table is given focus on tab', () => {
       browser.keys(['Tab']);
 
       expect(browser.$$('button:focus')).toBeElementsArrayOfSize(1);
       Terra.validates.element('table-focus-first-button', { columnResizeSelector: focusElementSelector });
     });
 
-    it('validates that the proper grid element is selected when Tab is used to give initial focus to the grid', () => {
+    it('validates that the second interactive element in the table is given focus with two tabs', () => {
       browser.keys(['Tab', 'Tab']);
 
       expect(browser.$$('input:focus')).toBeElementsArrayOfSize(1);