+++ packages/terra-framework-docs/CHANGELOG.md
@@ -5,6 +5,7 @@
 * Added
   * Added test to cover enabling zebra striping for the `terra-table` component.
   * Added test for tab focus for scrollable tables for the `terra-table` component.
+  * Added examples and tests for table sections in the `terra-table` component.
 
 * Updated
   * Removed columnHeaderHeight from tests using the default implementation for `terra-data-grid`.