+++ packages/terra-framework-docs/CHANGELOG.md
@@ -7,7 +7,7 @@
 * Added
   * Added test to cover enabling zebra striping for the `terra-table` component.
   * Added test for tab focus for scrollable tables for the `terra-table` component.
-  * Added examples and tests to cover the behavior of interactive elements within a table for `terra=table`.
+  * Added examples and tests to cover the behavior of interactive elements within a table for `terra-table`.
 
 * Updated
   * Removed columnHeaderHeight from tests using the default implementation for `terra-data-grid`.