+++ packages/terra-framework-docs/CHANGELOG.md
@@ -12,12 +12,10 @@
   * Added test to cover enabling zebra striping for the `terra-table` component.
   * Added test for tab focus for scrollable tables for the `terra-table` component.
   * Added examples and tests to cover the behavior of interactive elements within a table for `terra-table`.
+  * Added example for Flowsheet Data Grid containing missing results.
 
 * Updated
   * Removed columnHeaderHeight from tests using the default implementation for `terra-data-grid`.
-
-* Changed
-  * Updated test for FlowsheetDataGrid in `terra-data-grid`.
 
 ## 1.41.0 - (October 17, 2023)