+++ packages/terra-framework-docs/CHANGELOG.md
@@ -6,7 +6,7 @@
   * Fixed the `terra-tabs` icon-only examples keyboard navigation issue.
 
 * Added
-  * Added example for Flowsheet Data Grid containing missing results.
+  * Added example for Flowsheet Data Grid containing "No Result" cells.
 
 ## 1.42.0 - (October 20, 2023)